feat: Update civitai.js to handle HTTP error responses
The `civitai.js` file has been updated to handle HTTP error responses. Previously, when a response was not OK, an error was thrown with the status code. Now, the error response body is also included in the error message. This provides more information about the error and helps with debugging.